add fixes based off review
