+++ cli/command/deploy_test.go
@@ -1,8 +1,6 @@
 package command
 
 import (
-	"io/ioutil"
-	"os"
 	"testing"
 
 	"github.com/quintilesims/layer0/common/models"
@@ -182,17 +180,4 @@ func TestFilterDeploySummaries(t *testing.T) {
 	assert.Contains(t, output, input[2])
 	assert.Contains(t, output, input[5])
 	assert.Contains(t, output, input[8])
-}
-
-func createTempFile(t *testing.T, content string) (*os.File, func()) {
-	file, err := ioutil.TempFile(os.TempDir(), "")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if _, err := file.Write([]byte(content)); err != nil {
-		t.Fatal(err)
-	}
-
-	return file, func() { os.Remove(file.Name()) }
 }
